% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{add_behdata}
\alias{add_behdata}
\title{Merge behavioral data with itrackR object}
\usage{
add_behdata(obj, beh, append = FALSE)
}
\arguments{
\item{obj}{an itrackR object}

\item{beh}{a data frame containing behavioral data. Should at least have the column \code{ID} which matches ID names/numbers in \code{obj$subs}.}

\item{append}{set to TRUE if you've already added behavioral data and you're adding more columns (default = FALSE, overwrite existing data).}
}
\value{
data frame is saved in \code{obj$beh}, with the extra column \code{eyetrial} which matches the eyetracking data.
Variables in \code{obj$beh} can be used for subsetting when calling \code{\link{eyemerge}}, \code{\link{plot.itrackR}}, and \code{\link{fixation_timeseries}}.
}
\description{
Take a data frame containing behaivoral data and merge with the eyetracking data in an itrackR object. Merging is done based on \code{ID} and any
index variables set using \code{\link{set_index}}. Index variables should match the same ones that are in \code{obj$header}.
}
\examples{
\dontrun{
# itrackr.data('edfs') returns full path to 2 edf files
z <- itrackr(edfs=itrackr.data('edfs'))

#Extracts messages like "BLOCK 1", and "TRIAL 7" and creates variables "Block" and "Trial", giving them values
# 1 and 7.
z <- set_index(z,c('Block', 'Trial'), patterns=c('BLOCK [0-9]*', 'TRIAL [0-9][0-9]*), numeric.only = TRUE)

beh <- itrackr.data('beh)

z <- add_behdata(z,beh)

#notice the new eyetrial variable...
View(z$beh)

}


}
\seealso{
\code{\link{eyemerge}}
}

