% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{eyemerge}
\alias{eyemerge}
\title{Create data frame of merged eyetracking and behavioral data}
\usage{
eyemerge(obj, eyedata = "fixations", behdata = "all", all.rois = F,
  event = NULL, roi = NULL, trialtime = TRUE, condition = NULL,
  condition.str = FALSE)
}
\arguments{
\item{obj}{an itrackR object}

\item{eyedata}{type of eyetracking data to include. Can be:
\itemize{
 \item \code{'fixations'} (default). Fixation start/end times and x/y coordinates
 \item \code{'saccades'} Saccade start/end times and star/end x/y coordinates
 \item \code{'blinks'} Blink start/end times
 \item \code{'epoched_fixations'} fixation data epoched around some event using \code{\link{epoch_fixations}}
}}

\item{behdata}{list of variables from \code{obj$beh} to include in the output (default = 'all').}

\item{all.rois}{whether to include hits/misses for all rois in itrackR object (default = FALSE).}

\item{event}{the timelocking event when getting epoched fixations}

\item{roi}{the roi when getting epoched fixations}

\item{trialtime}{whether to convert fixation/saccade times relative to the current trial, or leave relative to whole experiment.}

\item{condition}{subset data according to some behavioral variables. Can use unquoted variable names as in \code{\link{subset}}. (e.g., condition = Block <= 5)}

\item{condition.str}{whether the condition is saved as a string ('Block <= 5'). Used internally.}
}
\value{
returns a data frame with the variables from \code{obj$beh} merged with the eyetracking data requested. Merging is based on \code{ID}, and \code{obj$indexvars}.
This corresponds to \code{obj$beh$eyetrial} as well.
}
\description{
Function for creating a flat table of eyetracking and behavioral data from itrackR object. Can be used for gathering fixation, saccade or blink data,
as well as fixation data epoched around some event. Will include all behavioral variables, or a subset of named variables.
}
\examples{
\dontrun{
# itrackr.data('edfs') returns full path to 2 edf files
z <- itrackr(edfs=itrackr.data('edfs'))

#Extracts messages like "BLOCK 1", and "TRIAL 7" and creates variables "Block" and "Trial", giving them values
# 1 and 7.
z <- set_index(z,c('Block', 'Trial'), patterns=c('BLOCK [0-9]*', 'TRIAL [0-9][0-9]*), numeric.only = TRUE)

beh <- itrackr.data('beh)

z <- add_behdata(z,beh)

#get fixation data
fixes <- eyemerge(z,'fixations')

#get data only when Task==2
fixes <- eyemerge(z, 'fixations', condition = Task==2)

#saccades
saccs <- eyemerge(z,'saccades')

}


}
\seealso{
\code{\link{epoch_fixations}} \code{\link{set_index}}
}

