% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{find_messages}
\alias{find_messages}
\title{Find messages sent to Eyelink}
\usage{
find_messages(obj, varnames, patterns, numeric.only = FALSE,
  timestamp = FALSE)
}
\arguments{
\item{obj}{an itrackR object}

\item{varnames}{a list specifying what we want the variable names to be called}

\item{patterns}{search strings for the messages of interest (can use regular expressions)}

\item{numeric.only}{return only numeric information from the messages (default = FALSE). Useful for indexing variables (e.g., "Trial 5" is converted to 5).}

\item{timestamp}{return the timestamp of the message, not the message itself (default = FAlSE)}
}
\value{
all messages are stored in \code{obj$header}, based on the \code{varnames} specified.
When calling \code{eyemerge}, these will be appended to the data frame.
}
\description{
General function for finding messages sent to Eyelink and extracting information from them. Useful if you want timestamps
of particular events merged with behavioral data. Also used by \code{\link{set_index}} to extract information for index variables (for merging).
}
\examples{
\dontrun{
# itrackr.data('edfs') returns full path to 2 edf files
z <- itrackr(edfs=itrackr.data('edfs'))

#Loads all edf
z <- find_messages(z,c("STIMONSET"), patterns=c('STIMONSET'), timestamp = TRUE)
}


}
\seealso{
\code{\link{set_index}}
}

