% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHeatmap.R
\name{iHeatmap}
\alias{iHeatmap}
\title{iHeatmap}
\usage{
iHeatmap(x, y, stim, data)
}
\arguments{
\item{x}{[TODO]}

\item{y}{[TODO]}

\item{stim}{[TODO]}

\item{data}{[TODO]}
}
\value{
[TODO]
}
\description{
iHeatmap
}
\examples{
[TODO]
}

