% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{itrackr}
\alias{itrackr}
\title{Create an itrackR object}
\usage{
itrackr(edfs = NULL, path = NULL, pattern = "*.edf",
  resolution = c(1024, 768), binocular = FALSE)
}
\arguments{
\item{edfs}{list of edf file names as strings. Not necessary if you use \code{path} and \code{pattern}}

\item{path}{if not providing edf file names, you can provide a path to a directory and a search string}

\item{pattern}{search string for finding edf files, as a regular expression (default = '*.edf')}

\item{resolution}{a list specifying the x and y resolution of the monitor (default = c(1024,768))}

\item{binocular}{whether the experiment used binocular tracking (currently not supported, default = FALSE)}
}
\description{
Create an itrackR object by loading edfs
}
\examples{
\dontrun{
# itrackr.data('edfs') returns full path to 2 edf files
z <- itrackr(edfs=itrackr.data('edfs'))

#Loads all edf files starting with 3 from a certain directory
z <- itrackr(path='/path/to/some/directory', pattern = '3*.edf')
}

}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}
}

