% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{load_edfs}
\alias{load_edfs}
\title{load edf files}
\usage{
load_edfs(obj, path = ".", pattern = "*.edf", recursive = FALSE)
}
\description{
Internal function used by \code{itrackr} to load edfs. Do not call directly.
}

