% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{load_txt}
\alias{load_txt}
\title{load from text file from Eyelink DataViewer reports}
\usage{
load_txt(obj, type = "fixations", filename, sep = "\\t")
}
\description{
Experimental function for loading data from text files created from reports in DataViewer. Currently unused.
}

