% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{makeROIs}
\alias{makeROIs}
\title{Create regions of interest (ROIs) for itrackR object}
\usage{
makeROIs(obj, coords = data.frame(x = c(0), y = c(0)), shapes = "circle",
  radius = 0, xradius = 0, yradius = 0, angles = NULL, names = NULL,
  polygon = data.frame(x = c(), y = c()), append = F)
}
\arguments{
\item{obj}{an itrackR object}

\item{coords}{data frame or matrix specifying x and y coordinates for the center of each ROI (default = (0,0), upper-left corner)}

\item{shapes}{shape for each ROI listed, can be \code{'circle'}, \code{'ellipse'}, or \code{'polygon'}}

\item{radius}{for circular ROIs, radius of the circle (in pixels)}

\item{xradius}{for elliptical ROIs, radius along the x dimension}

\item{yradius}{for elliptical ROIs, radius along the y dimension}

\item{angles}{for elliptical ROIs, rotation for each ROI in degrees.}

\item{names}{optional, name for each ROI. Defaults to 1 to number of ROIs listed. ROIs should not have duplicate names!}

\item{polygon}{for polygonal ROIS, data frame or matrix specifying corners of polygon}

\item{append}{whether we're adding to our list of ROIs or overwriting (default = FALSE, overwrite)}
}
\value{
ROI information is stored in \code{obj$rois}. Uses the \code{\link{spatstat}} package to store ROI information.
To get ROI information as a data frame, you can use the internal function \code{\link{roi2df}}. ROIs can be plotted using \code{\link{plot.rois}}.
}
\description{
Create regions of interest (ROIs) for an itrackR object. ROIs can be circular, elliptical, or polygons. After creating ROIs, you use \code{\link{calcHits}}
to determine whether each fixation/saccade fell within each ROI. The helper functions \code{\link{radialCoords}} and \code{\link{roiFlower}} can be used to
generate coordinates around some set of coordinates.
}
\examples{
\dontrun{
# itrackr.data('edfs') returns full path to 2 edf files
z <- itrackr(edfs=itrackr.data('edfs'))

coords <- matrix(c(512,384,100,100),nrow=2,byrow=T)

z <- makeROIs(z,coords,shape='ellipse',xradius=60, yradius=120, angles = c(45, 75))



}


}
\seealso{
\code{\link{plot.rois}} \code{\link{radialCoords}}  \code{\link{roiFlower}} \code{\link{calcHits}}
}

