% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.itrackR}
\alias{plot.itrackR}
\title{Plot data from itrackR object}
\usage{
\method{plot}{itrackR}(obj, zoom = TRUE, crosshairs = TRUE, rois = TRUE,
  whichROIs = "all", names = FALSE, IDs = c(), summarize = 0,
  quick = F, condition = NULL, oneplot = F)
}
\arguments{
\item{obj}{the itrackR object to plot}

\item{zoom}{whether to zoom within the bounds of the screen resolution (obj$resolution)}

\item{rois}{show outlines of the rois in the plots (default = TRUE)}

\item{whichROIs}{a list of ROI names if you want to display only a subset of them}

\item{IDs}{list of subject IDs if you want to plot only a subset of subjects}

\item{summarize}{create 2d histogram of fixations. Specify the number of bins.}

\item{condition}{for subsetting based on behavioral data. Syntax is the same as used in \code{subset}}

\item{oneplot}{whether to plot all subjects' data on a single plot (default = FALSE). By default it creates a separate plot per subject}

\item{corsshairs}{show dotted red lines crossing through the center of the screen (default = TRUE)}
}
\description{
This creates a scatterplot of fixation data from an itrackR object
}
\examples{
\dontrun{
z <- itrackr(edfs=itrackr.data('edfs'))
plot(z, zoom = T)
}

}
\author{
Jason Hubbard, \email{hubbard3@uoregon.edu}
}

