% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{radialCoords}
\alias{radialCoords}
\title{Generates a set of "numpoints" coordinates around a circle with specified radius,
centered around x and y, and (optionally) with at starting_angle (default=0 degrees)}
\usage{
radialCoords(x, y, numpoints, radius, starting_angle = 0)
}
\description{
Generates a set of "numpoints" coordinates around a circle with specified radius,
centered around x and y, and (optionally) with at starting_angle (default=0 degrees)
}

