% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{set_index}
\alias{set_index}
\title{Find messages to use as index variables in itrackR object}
\usage{
set_index(obj, varnames, patterns, numeric.only = FALSE)
}
\arguments{
\item{obj}{an itrackR object}

\item{varnames}{a list specifying what we want the variable names to be called}

\item{patterns}{search strings for the messages of interest (can use regular expressions)}

\item{numeric.only}{return only numeric information from the messages (default = FALSE). Useful for indexing variables (e.g., "Trial 5" is converted to 5).}
}
\value{
all variables are stored in \code{obj$header}, based on the \code{varnames} specified. You can also see them in \code{obj$indexvars}.
When calling \code{eyemerge}, these will be appended to the data frame. When using \code{\link{add_behdata}}, variable names
in the behavioral file should match \code{varnames}.
}
\description{
Mostly a wrapper for \link{\code{find_messages}}, but extracts messages and sets them as \code{obj$indexvars}. These are used in other functions
like \link{\code{add_behdata}} and \link{\code{eyemerge}} to use as index variables for merging with behavioral data. This assumes you sent messages
to Eyelink during an experiment to specify the trial number. It can also handle multiple variables (e.g., Block 5, Trial 6).
}
\examples{
\dontrun{
# itrackr.data('edfs') returns full path to 2 edf files
z <- itrackr(edfs=itrackr.data('edfs'))

#Extracts messages like "BLOCK 1", and "TRIAL 7" and creates variables "Block" and "Trial", giving them values
# 1 and 7.
z <- set_index(z,c('Block', 'Trial'), patterns=c('BLOCK [0-9]*', 'TRIAL [0-9][0-9]*'), numeric.only = TRUE)

#you can see the result by looking at the header
View(z$header)
}


}
\seealso{
\code{\link{add_behdata}}, \code{\link{find_messages}}
}

