% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itrackR.R
\name{undrift}
\alias{undrift}
\title{Reverse drift correction performed on an itrackR object}
\usage{
undrift(obj)
}
\arguments{
\item{obj}{an itrackR object}
}
\value{
All fixations in \code{obj$fixations} are adjusted back to original coordinates. \code{obj$transform} is deleted.
}
\description{
In case you want to undo the drift correction performed by \code{\link{drift_correct}}. It puts fixations back to original coordinates and deletes \code{obj$transform}.
}
\examples{
\dontrun{
# itrackr.data('edfs') returns full path to 2 edf files
z <- itrackr(edfs=itrackr.data('edfs'))

#Extracts messages like "BLOCK 1", and "TRIAL 7" and creates variables "Block" and "Trial", giving them values
# 1 and 7.
z <- set_index(z,c('Block', 'Trial'), patterns=c('BLOCK [0-9]*', 'TRIAL [0-9][0-9]*), numeric.only = TRUE)

beh <- itrackr.data('beh)

z <- add_behdata(z,beh)

#fixations before correcting
plot(z)

#perform drift correction for each subject and Block.
z <- drift_correct(z, vars=c('Block'))

#after correcting
plot(z)


#never mind
z <- undrift(z)

#back to normal
plot(z)

}


}
\seealso{
\code{\link{drift_correct}}
}

